Rotation Fix: Minor bugfix supporting floating point Viewboxes

+++ src/scratchtocatrobat/tools/svgtopng.py
@@ -365,8 +365,7 @@ def _get_viewbox_values(view_box_str):
     
     st = StringTokenizer(view_box_str," ")
     while st.hasMoreTokens():
-        view_box_values.append(int(st.nextToken()))  
-     
+        view_box_values.append(float(st.nextToken()))  
     return view_box_values
